LayoutTests/http/tests/w3c/dom/events/Event-initEvent-expected.txt


FAIL Properties of initEvent(type, true, true) assert_equals: isTrusted expected (boolean) false but got (undefined) undefined
FAIL Properties of initEvent(type, true, false) assert_equals: isTrusted expected (boolean) false but got (undefined) undefined
FAIL Properties of initEvent(type, false, true) assert_equals: isTrusted expected (boolean) false but got (undefined) undefined
FAIL Properties of initEvent(type, false, false) assert_equals: isTrusted expected (boolean) false but got (undefined) undefined
PASS Calling initEvent multiple times (getting type). 
PASS Calling initEvent multiple times (not getting type). 
PASS Calling initEvent must not have an effect during dispatching. 
PASS Calling initEvent must unset the stop propagation flag. 
PASS Calling initEvent during propagation. 
